+++ src/objtools/writers/writedb/writedb_convert.cpp
@@ -258,7 +258,7 @@ public:
             }
         }
         
-        int num_amb = m_Regions.size();
+        int num_amb = (int) m_Regions.size();
         
         // The size packed here is actually the number of *words* used
         // rather than the number of ambiguous regions; the new format
@@ -526,7 +526,7 @@ void WriteDB_Ncbi4naToBinary(const CSeq_inst & seqinst,
     const vector<char> & na4 = seqinst.GetSeq_data().GetNcbi4na().Get();
     int base_length = seqinst.GetLength();
     
-    WriteDB_Ncbi4naToBinary(& na4[0], na4.size(), base_length, seq, amb);
+    WriteDB_Ncbi4naToBinary(& na4[0], (int) na4.size(), base_length, seq, amb);
 }
 
 void WriteDB_StdaaToBinary(const CSeq_inst & si, string & seq)
@@ -548,7 +548,7 @@ void WriteDB_EaaToBinary(const CSeq_inst & si, string & seq)
     CSeqConvert::Convert(v,
                          CSeqUtil::e_Ncbieaa,
                          0,
-                         v.size(),
+                         (int) v.size(),
                          seq,
                          CSeqUtil::e_Ncbistdaa);
 }
@@ -563,7 +563,7 @@ void WriteDB_IupacaaToBinary(const CSeq_inst & si, string & seq)
     CSeqConvert::Convert(v,
                          CSeqUtil::e_Iupacaa,
                          0,
-                         v.size(),
+                         (int) v.size(),
                          seq,
                          CSeqUtil::e_Ncbistdaa);
 }
